requirements.txt
appdirs==1.4.4
APScheduler==3.10.4
Backtesting==0.3.3
beautifulsoup4==4.12.2
binance-connector==3.5.1
bokeh==3.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.46.0
frozendict==2.3.10
html5lib==1.1
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.2
mplfinance==0.12.10b0
multitasking==0.0.11
numpy==1.26.2
packaging==23.2
pandas==2.1.3
peewee==3.17.0
Pillow==10.1.0
plotly==5.18.0
pybit==5.6.2
pycryptodome==3.19.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tenacity==8.2.3
tornado==6.4
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
xyzservices==2023.10.1
yfinance==0.2.33
